Converted to JavaFX properties

+++ Scheduler/src/model/User.java
@@ -6,5 +6,5 @@ package model;
  */
 public interface User extends Record {
     String getUserName();
-    short getActive();
+    int getActive();
 }
